requirements.txt
Django==4.2.3
psycopg2==2.9.6
celery==5.3.1
django-lifecycle==1.0.0
django-environ==0.10.0
gunicorn==20.1.0
requests==2.31.0
factory-boy==3.2.1
Faker==10.0.0
uvicorn==0.21.0
django-celery-beat==2.5.0
beautifulsoup4==4.11.1
lxml==4.9.2
django-import-export==3.3.1
dataclasses-json==0.6.1
django-multiselectfield==0.1.12
pandas==2.1.1
django-admin-rangefilter==0.11.2
